Split metrics to separate file and use constants for label
